--- a/.github/workflows/test.yaml
+++ b/.github/workflows/test.yaml
@@ -26,7 +26,7 @@ jobs:
           npm install
       - name: Run tests
         run: |
-          npm run test
+          npm run test -- --silent --maxWorkers=4
   test-component:
     needs: test-unit
     runs-on: ubuntu-latest
@@ -41,7 +41,7 @@ jobs:
           npm install
       - name: Run tests
         run: |
-          npm run test-component
+          npm run test-component -- --silent --maxWorkers=4
   test-e2e:
     needs: test-component
     runs-on: ubuntu-latest
@@ -56,4 +56,4 @@ jobs:
           npm install
       - name: Run tests
         run: |
-          npm run start & npm run test-e2e
+          npm run start & npm run test-e2e -- --silent --maxWorkers=4