--- a/cypress.config.js
+++ b/cypress.config.js
@@ -6,6 +6,7 @@ module.exports = defineConfig({
       // implement node event listeners here
     },
     baseUrl: "http://localhost:3000",
+    numTestsKeptInMemory: 4,
   },
 
   component: {