--- a/package.json
+++ b/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "vorgabe",
+  "name": "m450m324biztrips",
   "version": "0.1.0",
   "private": true,
   "dependencies": {