--- a/package.json
+++ b/package.json
@@ -24,8 +24,8 @@
     "start-api": "json-server --port 3001 --watch db.json --delay 1500",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "test-e2e": "npx cypress run --parallel --record",
-    "test-component": "npx cypress run --component --component --parallel --record",
+    "test-e2e": "npx cypress run --parallel",
+    "test-component": "npx cypress run --component --component --parallel",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {