| Optimization | Unit Tests | Component Tests | E2e Tests | Total |
| --- | ---- | ---- | ---- | ---- |
| None | 58s | 1m 30s | 1m 9s | **3m 37s** |
| Increased Workers (to 4) for Unit Tests |1m 15s | 1m 42s | 1m 33s | **4m 30s** |
| Increased Workers (to 2) for Unit Tests |1m 12s | 1m 29s | 1m 43s | **4m 24s** |
| Run Jobs in parallel | 1m 28s | 1m 27s | 1m 18s | **1m 28s** |

Sadly couldn't pass parallel flag into `cypress`:

```
You passed the --ci-build-id, --group, --tag, --parallel, or --auto-cancel-after-failures flag without also passing the --record flag.

The --parallel flag you passed was: true
```